chore: update V1 modal's folder to _deprecated
