+++ apps/ledger-live-desktop/src/renderer/families/evm/StakeFlowModal/EthStakingModalBody.tsx
@@ -1,55 +1,38 @@
-import { Flex, Text } from "@ledgerhq/react-ui";
-import { Account } from "@ledgerhq/types-live";
-import React, { useCallback, useState } from "react";
-import { useHistory } from "react-router-dom";
-import BigNumber from "bignumber.js";
-import { useTranslation } from "react-i18next";
-
-import { appendQueryParamsToDappURL } from "@ledgerhq/live-common/platform/utils/appendQueryParamsToDappURL";
-import { getCryptoCurrencyById } from "@ledgerhq/live-common/currencies/index";
 import { LiveAppManifest } from "@ledgerhq/live-common/platform/types";
+import { appendQueryParamsToDappURL } from "@ledgerhq/live-common/platform/utils/appendQueryParamsToDappURL";
+import { Flex } from "@ledgerhq/react-ui";
+import { Account, EthStakingProvider } from "@ledgerhq/types-live";
+import React, { useCallback, useState } from "react";
+import { useTranslation } from "react-i18next";
+import { useHistory } from "react-router-dom";
 import { track } from "~/renderer/analytics/segment";
 import CheckBox from "~/renderer/components/CheckBox";
-import EthStakeIllustration from "~/renderer/icons/EthStakeIllustration";
-
 import {
   CheckBoxContainer,
   LOCAL_STORAGE_KEY_PREFIX,
 } from "~/renderer/modals/Receive/steps/StepReceiveStakingFlow";
-import { ListProvider, ListProviders } from "./types";
 import { getTrackProperties } from "./utils/getTrackProperties";
 
 import ProviderItem from "./component/ProviderItem";
 
-const ethMagnitude = getCryptoCurrencyById("ethereum").units[0].magnitude;
-
-const ETH_LIMIT = BigNumber(32).times(BigNumber(10).pow(ethMagnitude));
-
-// Comparison fns for sorting providers by minimum ETH required
-const ascending = (a: ListProvider, b: ListProvider) => (a?.min || 0) - (b?.min || 0);
-const descending = (a: ListProvider, b: ListProvider) => (b?.min || 0) - (a?.min || 0);
-
 type Props = {
   account: Account;
-  singleProviderRedirectMode?: boolean;
   onClose?: () => void;
   hasCheckbox?: boolean;
   source?: string;
-  listProviders?: ListProviders;
+  providers: EthStakingProvider[];
 };
 
 export type StakeOnClickProps = {
-  provider: ListProvider;
+  provider: EthStakingProvider;
   manifest: LiveAppManifest;
 };
-
 export function EthStakingModalBody({
   hasCheckbox = false,
-  singleProviderRedirectMode = true,
   source,
   onClose,
   account,
-  listProviders = [],
+  providers,
 }: Props) {
   const { t } = useTranslation();
   const history = useHistory();
@@ -66,6 +49,7 @@ export function EthStakingModalBody({
         button: providerConfigID,
         ...getTrackProperties({ value, modal: source }),
       });
+
       history.push({
         pathname: value,
         ...(customDappUrl ? { customDappUrl } : {}),
@@ -78,15 +62,6 @@ export function EthStakingModalBody({
     [history, account.id, onClose, source],
   );
 
-  const redirectIfOnlyProvider = useCallback(
-    (stakeOnClickProps: StakeOnClickProps) => {
-      if (singleProviderRedirectMode && listProviders.length === 1) {
-        stakeOnClick(stakeOnClickProps);
-      }
-    },
-    [singleProviderRedirectMode, listProviders.length, stakeOnClick],
-  );
-
   const checkBoxOnChange = useCallback(() => {
     const value = !doNotShowAgain;
     global.localStorage.setItem(`${LOCAL_STORAGE_KEY_PREFIX}${account?.currency?.id}`, `${value}`);
@@ -97,50 +72,28 @@ export function EthStakingModalBody({
     });
   }, [doNotShowAgain, account?.currency?.id, source]);
 
-  const hasMinValidatorEth = account.spendableBalance.isGreaterThan(ETH_LIMIT);
-
-  const listProvidersSorted = listProviders.sort(hasMinValidatorEth ? descending : ascending);
-
   return (
-    <Flex flexDirection={"column"} alignItems="center" width={"100%"}>
-      <Flex flexDirection="column" alignItems="center" rowGap={16}>
-        <Text ff="Inter|SemiBold" fontSize="24px" lineHeight="32px">
-          {t("ethereum.stake_v2.title")}
-        </Text>
-        {listProviders.length <= 1 && (
-          <Flex justifyContent="center" py={20} width="100%">
-            <EthStakeIllustration size={140} />
+    <Flex flexDirection="column" width="100%">
+      <Flex flexDirection="column" rowGap={2}>
+        {providers.map(x => (
+          <Flex key={x.id} flexDirection="column">
+            <ProviderItem provider={x} stakeOnClick={stakeOnClick} />
           </Flex>
-        )}
-        <Text textAlign="center" color="neutral.c70" fontSize={14} maxWidth={360}>
-          {t("ethereum.stake.subTitle")}
-        </Text>
+        ))}
       </Flex>
-      <Flex flexDirection={"column"} mt={5} px={20} width="100%">
-        <Flex flexDirection={"column"} width="100%">
-          {listProvidersSorted.map(item => (
-            <Flex key={item.id} width="100%" flexDirection={"column"}>
-              <ProviderItem
-                provider={item}
-                stakeOnClick={stakeOnClick}
-                redirectIfOnlyProvider={redirectIfOnlyProvider}
-              />
-            </Flex>
-          ))}
-        </Flex>
-        {hasCheckbox && (
-          <CheckBoxContainer
-            p={3}
-            borderRadius={8}
-            borderWidth={0}
-            width={"100%"}
-            onClick={checkBoxOnChange}
-            mt={15}
-          >
-            <CheckBox isChecked={doNotShowAgain} label={t("receive.steps.staking.notShow")} />
-          </CheckBoxContainer>
-        )}
-      </Flex>
+      {/* TODO review */}
+      {hasCheckbox && (
+        <CheckBoxContainer
+          p={3}
+          borderRadius={8}
+          borderWidth={0}
+          width="100%"
+          onClick={checkBoxOnChange}
+          mt={15}
+        >
+          <CheckBox isChecked={doNotShowAgain} label={t("receive.steps.staking.notShow")} />
+        </CheckBoxContainer>
+      )}
     </Flex>
   );
 }
