+++ apps/ledger-live-desktop/src/renderer/families/evm/StakeFlowModal/component/ProviderItem.tsx
@@ -1,12 +1,61 @@
 import { useRemoteLiveAppManifest } from "@ledgerhq/live-common/platform/providers/RemoteLiveAppProvider/index";
-import { Flex, Icon, Tag as TagCore, Text } from "@ledgerhq/react-ui";
-import React, { useCallback, useEffect, useMemo } from "react";
+import { useLocalLiveAppManifest } from "@ledgerhq/live-common/wallet-api/LocalLiveAppProvider/index";
+import { CryptoIcon, Flex, Icon, Text } from "@ledgerhq/react-ui";
+import { EthStakingProvider } from "@ledgerhq/types-live";
+import React, { useCallback, useMemo } from "react";
 import { useTranslation } from "react-i18next";
 import styled, { DefaultTheme, StyledComponent } from "styled-components";
+import ProviderIcon from "~/renderer/components/ProviderIcon";
 import { StakeOnClickProps } from "../EthStakingModalBody";
-import { StakingIcon } from "../StakingIcon";
-import { ListProvider } from "../types";
-import { useLocalLiveAppManifest } from "@ledgerhq/live-common/wallet-api/LocalLiveAppProvider/index";
+
+const IconContainer = styled.div(
+  ({ theme }) => `
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: ${theme.space[6]}px;
+  height: ${theme.space[6]}px;
+  border-radius: 100%;
+  background-color: ${theme.colors.opacityDefault.c05};
+  margin-top: ${theme.space[3]}px;
+`,
+);
+
+const ProviderIconContainer = styled.div(
+  () => `
+  display: flex;
+`,
+);
+
+export function StakingIcon({ icon }: { icon?: string }) {
+  if (!icon) {
+    return null;
+  }
+
+  const [iconName, iconType] = icon.split(":");
+
+  // if no icon type then treat as "normal" icon.
+  if (!iconType) {
+    return (
+      <IconContainer>
+        <Icon name={iconName} size={14} />
+      </IconContainer>
+    );
+  }
+
+  if (iconType === "crypto") {
+    return <CryptoIcon name={iconName} size={40} />;
+  }
+  if (iconType === "provider") {
+    return (
+      <ProviderIconContainer>
+        <ProviderIcon name={iconName} size="L" />
+      </ProviderIconContainer>
+    );
+  }
+
+  return null;
+}
 
 export const Container: StyledComponent<
   "div",
@@ -15,77 +64,59 @@ export const Container: StyledComponent<
   never
 > = styled(Flex)`
   cursor: pointer;
-  border-radius: 8px;
+  background-color: ${p => p.theme.colors.opacityDefault.c05};
   :hover {
     background-color: ${p => p.theme.colors.primary.c10};
   }
 `;
 
-export const Tag = styled(TagCore)`
-  padding: 3px 6px;
-  > span {
-    font-size: 11px;
-    text-transform: none;
-    font-weight: bold;
-    line-height: 11.66px;
-  }
-`;
-
 type Props = {
-  provider: ListProvider;
+  provider: EthStakingProvider;
   stakeOnClick(_: StakeOnClickProps): void;
-  redirectIfOnlyProvider(_: StakeOnClickProps): void;
 };
 
-const ProviderItem = ({ provider, stakeOnClick, redirectIfOnlyProvider }: Props) => {
-  const { t, i18n } = useTranslation();
+const ProviderItem = ({ provider, stakeOnClick }: Props) => {
+  const { t } = useTranslation();
 
   const localManifest = useLocalLiveAppManifest(provider.liveAppId);
   const remoteManifest = useRemoteLiveAppManifest(provider.liveAppId);
 
   const manifest = useMemo(() => remoteManifest || localManifest, [localManifest, remoteManifest]);
 
-  const hasTag = !!provider?.min && i18n.exists(`ethereum.stake.${provider.id}.tag`);
-
-  useEffect(() => {
-    if (manifest) redirectIfOnlyProvider({ provider, manifest });
-  }, [redirectIfOnlyProvider, provider, manifest]);
-
-  const stakeLink = useCallback(() => {
-    if (manifest) stakeOnClick({ provider, manifest });
+  const handleClick = useCallback(() => {
+    if (manifest) {
+      stakeOnClick({ provider, manifest });
+    }
   }, [provider, stakeOnClick, manifest]);
 
   return (
     <Container
-      pl={3}
-      onClick={stakeLink}
-      py={4}
+      p={2}
+      borderRadius={2}
+      onClick={handleClick}
+      alignItems="center"
+      columnGap={2}
       data-testid={`stake-provider-container-${provider.id}`}
     >
       <StakingIcon icon={provider.icon} />
-      <Flex flexDirection={"column"} ml={5} flex={"auto"} alignItems="flex-start">
-        <Flex alignItems="center">
-          <Text variant="bodyLineHeight" fontSize={14} fontWeight="semiBold" mr={2}>
-            {t(`ethereum.stake.${provider.id}.title`)}
-          </Text>
-          {hasTag && (
-            <Tag
-              size="small"
-              active
-              type="plain"
-              style={{ fontFamily: "14px", textTransform: "none" }}
-            >
-              {t(`ethereum.stake.${provider.id}.tag`)}
-            </Tag>
-          )}
-        </Flex>
-
+      <Flex alignItems="flex-start" flex={2} flexDirection="column">
+        <Text variant="bodyLineHeight" fontSize={14} fontWeight="semiBold" mr={2}>
+          {t(`ethereum.stake.provider.${provider.id}.title`)}
+        </Text>
         <Text variant="paragraph" fontSize={13} color="neutral.c70">
-          {t(`ethereum.stake.${provider.id}.description`)}
+          {provider.lst
+            ? t("ethereum.stake.lst")
+            : provider.min
+              ? t("ethereum.stake.required_minimum", {
+                  min: provider.min,
+                })
+              : t("ethereum.stake.no_minimum")}
         </Text>
       </Flex>
-      <Flex width={40} justifyContent="center" alignItems="center">
-        <Icon name="ChevronRight" size={25} />
+      <Flex flex={1} flexWrap="wrap" justifyContent="right">
+        <Text variant="paragraph" fontSize={13} color="neutral.c70" textAlign="right">
+          {t(`ethereum.stake.rewards_strategy.${provider.rewardsStrategy}`)}
+        </Text>
       </Flex>
     </Container>
   );
