+++ apps/ledger-live-desktop/src/renderer/families/evm/StakeFlowModal_deprecated/utils/getTrackProperties.ts
