+++ apps/ledger-live-desktop/static/i18n/en/app.json
@@ -3698,7 +3698,7 @@
     }
   },
   "ethereum": {
-    "stake": {
+    "stake_deprecated": {
       "title": "Grow your ETH holdings",
       "subTitle": "Stake your ETH and grow your holdings by receiving daily rewards directly into your Ledger Live.",
       "tagMinimum": "Requires {{min}} ETH",
@@ -3730,7 +3730,7 @@
         "supportLink": "Learn more about Figment validator"
       }
     },
-    "stake_v2": {
+    "stake": {
       "title": "Select staking provider",
       "category": {
         "all": {
