+++ apps/ledger-live-mobile/src/families/evm/StakingDrawer/EvmStakingDrawerBody.tsx
@@ -1,23 +1,20 @@
-import { Flex, Text } from "@ledgerhq/native-ui";
-import React, { useCallback } from "react";
-import { useTranslation } from "react-i18next";
-
-import { ParamListBase, useNavigation } from "@react-navigation/native";
 import { LiveAppManifest } from "@ledgerhq/live-common/platform/types";
 import { appendQueryParamsToDappURL } from "@ledgerhq/live-common/platform/utils/appendQueryParamsToDappURL";
-
-import { ListProvider } from "./types";
-import { EvmStakingDrawerProvider } from "./EvmStakingDrawerProvider";
+import { Flex } from "@ledgerhq/native-ui";
+import { ParamListBase, useNavigation } from "@react-navigation/native";
+import { StackNavigationProp } from "@react-navigation/stack";
+import React, { useCallback } from "react";
 import { useAnalytics } from "~/analytics";
 import { NavigatorName, ScreenName } from "~/const";
-import { StackNavigationProp } from "@react-navigation/stack";
+import { EvmStakingDrawerProvider } from "./EvmStakingDrawerProvider";
+import { ListProvider } from "./type";
 
-type Props = {
+interface Props {
   providers: ListProvider[];
   singleProviderRedirectMode: boolean;
   accountId: string;
   onClose(callback: () => void): void;
-};
+}
 
 export function EvmStakingDrawerBody({
   providers,
@@ -25,7 +22,6 @@ export function EvmStakingDrawerBody({
   accountId,
   onClose,
 }: Props) {
-  const { t } = useTranslation();
   const navigation = useNavigation<StackNavigationProp<ParamListBase, string, NavigatorName>>();
 
   const { track, page } = useAnalytics();
@@ -63,23 +59,15 @@ export function EvmStakingDrawerBody({
   );
 
   return (
-    <Flex rowGap={16}>
-      <Flex rowGap={24} pb={8}>
-        <Text variant="h4">{t("stake.ethereum.title")}</Text>
-        <Text variant="body" lineHeight="21px" color="neutral.c70">
-          {t("stake.ethereum.subTitle")}
-        </Text>
-      </Flex>
-      <Flex rowGap={32} pb={32}>
-        {providers.map(provider => (
-          <EvmStakingDrawerProvider
-            key={provider.id}
-            provider={provider}
-            onProviderPress={onProviderPress}
-            redirectIfOneProvider={redirectIfOneProvider}
-          />
-        ))}
-      </Flex>
+    <Flex rowGap={16} pb={32}>
+      {providers.map(x => (
+        <EvmStakingDrawerProvider
+          key={x.id}
+          provider={x}
+          onProviderPress={onProviderPress}
+          redirectIfOneProvider={redirectIfOneProvider}
+        />
+      ))}
     </Flex>
   );
 }
