+++ apps/ledger-live-mobile/src/families/evm/StakingDrawer/index.tsx
@@ -1,26 +1,30 @@
-import React, { useEffect } from "react";
-import { useTheme } from "styled-components/native";
-
 import { useFeature } from "@ledgerhq/live-common/featureFlags/index";
-import { Flex, ScrollListContainer } from "@ledgerhq/native-ui";
+import { Box, ChipTabs, Flex, Icons, ScrollListContainer, Text } from "@ledgerhq/native-ui";
+import { EthStakingProvider, EthStakingProviderCategory } from "@ledgerhq/types-live";
+import React, { useEffect, useMemo, useState } from "react";
+import { useTranslation } from "react-i18next";
+import { useTheme } from "styled-components/native";
 import { Track } from "~/analytics";
+import QueuedDrawer from "~/components/QueuedDrawer";
 import { useRootDrawerContext } from "~/context/RootDrawerContext";
 import { EvmStakingDrawerBody } from "./EvmStakingDrawerBody";
-import QueuedDrawer from "~/components/QueuedDrawer";
-
 import type { ListProvider } from "./types";
+import { Linking, TouchableOpacity } from "react-native";
+
+type Option = EthStakingProviderCategory | "all";
+const OPTION_VALUES: Option[] = ["all", "liquid", "protocol", "pooling", "restaking"] as const;
+
+const INFO_ICON_CONTAINER_SIZE = 40;
 
 // compare functions for sorting providers list based on minimum required stake
 const ascending = (a: ListProvider, b: ListProvider) => (a?.min || 0) - (b?.min || 0);
 const descending = (a: ListProvider, b: ListProvider) => (b?.min || 0) - (a?.min || 0);
 
 export function EvmStakingDrawer() {
-  const { isOpen, onModalHide, openDrawer, onClose, drawer } = useRootDrawerContext();
+  const { openDrawer, drawer } = useRootDrawerContext();
   const ethStakingProviders = useFeature("ethStakingProviders");
   const isStakingProvidersEnabled = ethStakingProviders?.enabled;
   const providers: ListProvider[] | undefined = ethStakingProviders?.params?.listProvider;
-
-  const { theme: themeName } = useTheme();
 
   useEffect(() => {
     if (isStakingProvidersEnabled || (providers ?? []).length > 0) {
@@ -28,31 +32,122 @@ export function EvmStakingDrawer() {
     }
   }, [isStakingProvidersEnabled, openDrawer, providers]);
 
-  if (!ethStakingProviders || drawer.id !== "EvmStakingDrawer") return null;
+  return !ethStakingProviders || drawer.id !== "EvmStakingDrawer" ? null : (
+    <Content
+      accountId={drawer.props.accountId}
+      has32Eth={drawer.props.has32Eth ?? false}
+      providers={ethStakingProviders?.params?.listProvider ?? []}
+      singleProviderRedirectMode={drawer.props.singleProviderRedirectMode ?? true}
+    />
+  );
+}
 
-  const has32Eth = drawer.props.has32Eth ?? false;
+interface Props {
+  accountId: string;
+  has32Eth: boolean;
+  providers: EthStakingProvider[];
+  singleProviderRedirectMode: boolean;
+}
 
-  const listProvidersSorted: ListProvider[] = ethStakingProviders.params!.listProvider.sort(
-    has32Eth ? descending : ascending,
+function Content({ accountId, has32Eth, providers, singleProviderRedirectMode }: Props) {
+  const { isOpen, onModalHide, onClose } = useRootDrawerContext();
+
+  const { t } = useTranslation();
+
+  const { theme: themeName, colors } = useTheme();
+
+  const [selectedIndex, setSelectedIndex] = useState(0);
+  const selected = OPTION_VALUES[selectedIndex];
+
+  const OPTION_LABELS = useMemo(
+    () => OPTION_VALUES.map(value => t(`stake.ethereum.category.${value}.name`)),
+    [t],
+  );
+
+  const listProvidersSorted: ListProvider[] = useMemo(
+    () =>
+      providers
+        .sort(has32Eth ? descending : ascending)
+        .filter(x => !x.disabled && (selected === "all" || selected === x.category)),
+    [has32Eth, providers, selected],
   );
 
   return (
     <QueuedDrawer isRequestingToBeOpened={isOpen} onClose={onClose} onModalHide={onModalHide}>
-      <Flex justifyContent={"center"}>
-        <Track onMount event="ETH Stake Modal" />
+      <Track onMount event="ETH Stake Modal" />
+      <Flex minHeight="90%" rowGap={24}>
+        <Flex rowGap={16} alignItems="center">
+          <Text variant="h4" textAlign="center">
+            {t("stake.ethereum.title")}
+          </Text>
+          <ChipTabs
+            labels={OPTION_LABELS}
+            size="small"
+            activeIndex={selectedIndex}
+            onChange={setSelectedIndex}
+          />
+          <Text variant="body" lineHeight="21px" color="neutral.c70" textAlign="center">
+            {t(`stake.ethereum.category.${selected}.description`)}
+          </Text>
+        </Flex>
         <ScrollListContainer
-          directionalLockEnabled
-          paddingX={4}
           alwaysBounceVertical={false}
+          directionalLockEnabled
+          flex={1}
           indicatorStyle={themeName === "dark" ? "white" : "default"}
         >
           <EvmStakingDrawerBody
             onClose={onClose}
-            singleProviderRedirectMode={drawer.props.singleProviderRedirectMode ?? true}
-            accountId={drawer.props.accountId}
+            singleProviderRedirectMode={singleProviderRedirectMode}
+            accountId={accountId}
             providers={listProvidersSorted}
           />
         </ScrollListContainer>
+        <Box backgroundColor={colors.background.drawer} bottom={0} left={0} right={0}>
+          <TouchableOpacity
+            style={{
+              alignItems: "center",
+              backgroundColor: colors.primary.c20,
+              borderRadius: 8,
+              display: "flex",
+              flexDirection: "row",
+              gap: 16,
+              padding: 16,
+            }}
+            onPress={() =>
+              Linking.openURL(
+                selected === "restaking"
+                  ? "https://www.ledger.com/academy/what-is-ethereum-restaking"
+                  : "https://www.ledger.com/academy/ethereum-staking-how-to-stake-eth",
+              )
+            }
+          >
+            <Flex flexShrink={0} alignItems="center" justifyContent="center">
+              <Flex
+                alignItems="center"
+                backgroundColor="primary.c70a025"
+                width={INFO_ICON_CONTAINER_SIZE}
+                height={INFO_ICON_CONTAINER_SIZE}
+                borderRadius={INFO_ICON_CONTAINER_SIZE / 2}
+                justifyContent="center"
+                p={2}
+              >
+                <Icons.BookGraduation size="M" color="primary.c80" />
+              </Flex>
+            </Flex>
+            <Box flex={1} justifyContent="space-between">
+              <Text fontSize={14}>{t("stake.ethereum.how_it_works")}</Text>
+              <Box>
+                <Text fontSize={13} color={colors.neutral.c80}>
+                  {t("stake.ethereum.read_about")}
+                </Text>
+              </Box>
+            </Box>
+            <Flex flexShrink={0} alignItems="center" justifyContent="center">
+              <Icons.ExternalLink size="S" color={colors.neutral.c80} />
+            </Flex>
+          </TouchableOpacity>
+        </Box>
       </Flex>
     </QueuedDrawer>
   );
