+++ apps/ledger-live-mobile/src/families/evm/StakingDrawer/type.ts
