+++ apps/ledger-live-mobile/src/families/evm/StakingDrawer_deprecated/EvmStakingDrawerProvider.tsx
@@ -1,14 +1,14 @@
 import React, { useCallback, useEffect } from "react";
-import { useTranslation } from "react-i18next";
 import { TouchableOpacity } from "react-native";
+import { useTranslation } from "react-i18next";
 
-import { useRemoteLiveAppManifest } from "@ledgerhq/live-common/platform/providers/RemoteLiveAppProvider/index";
+import { Flex, Icon, Text, Tag } from "@ledgerhq/native-ui";
 import { LiveAppManifest } from "@ledgerhq/live-common/platform/types";
 import { useLocalLiveAppManifest } from "@ledgerhq/live-common/wallet-api/LocalLiveAppProvider/index";
-import { Flex, Text } from "@ledgerhq/native-ui";
+import { useRemoteLiveAppManifest } from "@ledgerhq/live-common/platform/providers/RemoteLiveAppProvider/index";
 
 import { EvmStakingDrawerProviderIcon } from "./EvmStakingDrawerProviderIcon";
-import { ListProvider } from "./type";
+import { ListProvider } from "./types";
 
 type Props = {
   provider: ListProvider;
@@ -37,7 +37,9 @@ export function EvmStakingDrawerProvider({
   const remoteManifest = useRemoteLiveAppManifest(provider.liveAppId);
   const manifest = remoteManifest || localManifest;
 
-  const { t } = useTranslation();
+  const { t, i18n } = useTranslation();
+  const hasTag: boolean =
+    !!provider?.min && i18n.exists(`stake.ethereum_deprecated.providers.${provider.id}.tag`);
 
   const providerPress = useCallback(() => {
     if (manifest) {
@@ -53,45 +55,25 @@ export function EvmStakingDrawerProvider({
 
   return (
     <TouchableOpacity onPress={providerPress}>
-      <Flex
-        flexDirection="row"
-        columnGap={16}
-        alignItems="center"
-        borderRadius={2}
-        p={5}
-        backgroundColor="opacityDefault.c05"
-      >
+      <Flex flexDirection="row" columnGap={16}>
         <EvmStakingDrawerProviderIcon icon={provider.icon} />
-        <Flex rowGap={2} alignItems="flex-start" flex={2}>
-          <Flex flexDirection="column" flex={1} alignItems="flex-start">
-            <Text variant="bodyLineHeight" fontSize={14} fontWeight="semiBold" mr={2}>
-              {t(`stake.ethereum_v2.provider.${provider.id}.title`)}
+        <Flex rowGap={2} alignItems="flex-start" flex={1}>
+          <Flex flexDirection="row" columnGap={8} rowGap={8} mb={2}>
+            <Text variant="body" fontWeight="semiBold">
+              {t(`stake.ethereum_deprecated.providers.${provider.id}.title`)}
             </Text>
-            <Text variant="paragraph" fontSize={13} color="neutral.c70">
-              {provider.lst
-                ? t("stake.ethereum_v2.lst")
-                : provider.min
-                  ? t("stake.ethereum_v2.required_minimum", {
-                      min: provider.min,
-                    })
-                  : t("stake.ethereum_v2.no_minimum")}
+            {hasTag && (
+              <Tag type="color">{t(`stake.ethereum_deprecated.providers.${provider.id}.tag`)}</Tag>
+            )}
+          </Flex>
+          <Flex rowGap={12}>
+            <Text variant="paragraph" lineHeight="20px" color="neutral.c70">
+              {t(`stake.ethereum_deprecated.providers.${provider.id}.description`)}
             </Text>
           </Flex>
         </Flex>
-        <Flex flex={1}>
-          {provider.rewardsStrategy ? (
-            <Text
-              variant="paragraph"
-              fontSize={12}
-              color="neutral.c70"
-              flexWrap="wrap"
-              textAlign="right"
-            >
-              {t(`stake.ethereum_v2.rewards_strategy.${provider.rewardsStrategy}`)}
-            </Text>
-          ) : (
-            <></>
-          )}
+        <Flex alignSelf="center">
+          <Icon name="ChevronRight" size={32} color="neutral.c100" />
         </Flex>
       </Flex>
     </TouchableOpacity>
