+++ apps/ledger-live-mobile/src/families/evm/StakingDrawer_deprecated/types.ts
